requirements.txt
dotted-notation